Creacion CursorToWikipediaArtistMapper
Se creo la clase para mantener los mismos niveles de abstracción
